refactor(将字符串相关函数移动至单独的文件):
