+++ lib/notion/convertInnerUrl.js
@@ -1,7 +1,7 @@
 import { idToUuid } from 'notion-utils'
-import { checkStrIsNotionId } from '../utils'
 import { getLastPartOfUrl } from '@/lib/utils/url'
 import { isBrowser } from '@/lib/utils/envUtils'
+import { checkStrIsNotionId } from '@/lib/utils/stringUtils'
 
 /**
  * 处理页面内连接跳转:
