+++ lib/notion/getNotionPost.js
@@ -3,7 +3,7 @@ import { idToUuid } from 'notion-utils'
 import { defaultMapImageUrl } from 'react-notion-x'
 import formatDate from '../utils/formatDate'
 import { getPage } from './getPostBlocks'
-import { checkStrIsNotionId, checkStrIsUuid } from '@/lib/utils'
+import { checkStrIsNotionId, checkStrIsUuid } from '@/lib/utils/stringUtils'
 
 /**
  * 根据页面ID获取内容
