+++ lib/utils/post.js
@@ -1,7 +1,6 @@
 /**
  * 文章相关工具
  */
-import { checkStartWithHttp } from '.'
 import { getPostBlocks } from '@/lib/db/getSiteData'
 import { getPageTableOfContents } from '@/lib/notion/getPageTableOfContents'
 import { siteConfig } from '@/lib/config'
@@ -11,6 +10,7 @@ import { getAiSummary } from '@/lib/plugins/aiSummary'
 import BLOG from '@/blog.config'
 import { uploadDataToAlgolia } from '@/lib/plugins/algolia'
 import { countWords } from '@/lib/plugins/wordCount'
+import { checkStartWithHttp } from '@/lib/utils/stringUtils'
 
 /**
  * 获取文章的关联推荐文章列表，目前根据标签关联性筛选
