+++ lib/utils/url.js
@@ -1,5 +1,5 @@
-import { checkStartWithHttp } from '@/lib/utils/index'
 import { isBrowser } from '@/lib/utils/envUtils'
+import { checkStartWithHttp } from '@/lib/utils/stringUtils'
 
 /**
  * 诸如 article/https://test.com 等被错误拼接前缀的slug进行处理
