+++ middleware.ts
@@ -1,9 +1,9 @@
 import { clerkMiddleware, createRouteMatcher } from '@clerk/nextjs/server'
 import { NextRequest, NextResponse } from 'next/server'
-import { checkStrIsNotionId } from '@/lib/utils'
 import { idToUuid } from 'notion-utils'
 import BLOG from './blog.config'
 import { getLastPartOfUrl } from '@/lib/utils/url'
+import { checkStrIsNotionId } from '@/lib/utils/stringUtils'
 
 /**
  * Clerk 身份验证中间件
